requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3 
boto3==1.17.99        
botocore==1.20.99     
bs4==0.0.1
certifi==2020.12.5    
chardet==4.0.0        
dj-database-url==0.5.0
Django==3.2
django-appconf==1.0.4
django-crispy-forms==1.11.2
django-heroku==0.3.1
django-imagekit==4.0.2
django-storages==1.11.1
gunicorn==20.1.0
html5lib==1.1
idna==2.10
jmespath==0.10.0
pilkit==2.0
Pillow==8.2.0
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
s3transfer==0.4.2
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.5
webencodings==0.5.1
whitenoise==5.2.0